Split proto/mod.rs into multiple files
This code is pure code movement, with the exception of a few `pub`
modifiers and `use` statements.

Fix #9.
